backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-chroma>=0.0.1
langchain-community>=0.0.10
chromadb>=0.4.18
python-dotenv>=1.0.0
pydantic>=2.4.2
pypdf>=3.15.1
tiktoken>=0.5.1
pandas>=2.0.0
jq>=1.4.1
unstructured>=0.10.30
markdown>=3.4.3
